MEan Calculation Bug Fixed
Fixed a bunch of bugs.  Biggest is that that mean calculation was giving nans and not converging.  Turns out there was a divided by sin(theta/2) when it should have been a multiply.  

Also fixed a bunch of dimentionality errors.  Added 6D capability to the mean function.  

We need to add 6D input acceptance to the covariance function next.